220607 18:33 lr main algo optimal, a few display errors + norme to fix

+++ Linear_Regression/lr.c
@@ -6,7 +6,7 @@
 /*   By: alambert <alambert@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/05 14:35:14 by alambert          #+#    #+#             */
-/*   Updated: 2022/06/06 19:45:32 by alambert         ###   ########.fr       */
+/*   Updated: 2022/06/07 17:58:52 by alambert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,8 +27,8 @@ int	ft_exit(void *id[2])
 
 int	main(void)
 {
-	float		fdb[9][24];
-	float		fv[24];
+	long double	fdb[9][24];
+	long double	fv[24];
 	void		*id[2];
 	int			i;
 
@@ -38,50 +38,50 @@ int	main(void)
 	ft_dset(fdb, fv);
 	while (i < 24)
 	{
-		printf("fdb[km][%d] = %e\n", i, fdb[km][i]);
-		printf("fdb[price]][%d] = %e\n", i, fdb[price][i]);
-		printf("fdb[k][%d] = %e\n", i, fdb[k][i]);
-		printf("fdb[p][%d] = %e\n", i, fdb[p][i]);
-		printf("fdb[prod][%d] = %e\n", i, fdb[prod][i]);
-		printf("fdb[sqkm][%d] = %e\n", i, fdb[sqkm][i]);
-		printf("fdb[sqprice]][%d] = %e\n", i, fdb[sqprice][i]);
-		printf("fdb[sqxmgap]][%d] = %e\n", i, fdb[sqxmgap][i]);
-		printf("fdb[sqymgap]][%d] = %e\n", i, fdb[sqymgap][i]);
+		printf("fdb[km][%d] = %Lf\n", i, fdb[km][i]);
+		printf("fdb[price]][%d] = %Lf\n", i, fdb[price][i]);
+		printf("fdb[k][%d] = %Lf\n", i, fdb[k][i]);
+		printf("fdb[p][%d] = %Lf\n", i, fdb[p][i]);
+		printf("fdb[prod][%d] = %Lf\n", i, fdb[prod][i]);
+		printf("fdb[sqkm][%d] = %Lf\n", i, fdb[sqkm][i]);
+		printf("fdb[sqprice]][%d] = %Lf\n", i, fdb[sqprice][i]);
+		printf("fdb[sqxmgap]][%d] = %Lf\n", i, fdb[sqxmgap][i]);
+		printf("fdb[sqymgap]][%d] = %Lf\n", i, fdb[sqymgap][i]);
 		printf("\n");
 		i++;
 	}
 	printf("\n");
-	printf("fv[iteration_cut] = %e\n", fv[iteration_cut]);
-	printf("fv[sumkm] = %e\n", fv[sumkm]);
-	printf("fv[sumprice] = %e\n", fv[sumprice]);
-	printf("fv[sumprod] = %e\n", fv[sumprod]);
-	printf("fv[sumsqkm] = %e\n", fv[sumsqkm]);
-	printf("fv[sumsqprice] = %e\n", fv[sumsqprice]);
+	printf("fv[iteration_cut] = %Lf\n", fv[iteration_cut]);
+	printf("fv[sumkm] = %Lf\n", fv[sumkm]);
+	printf("fv[sumprice] = %Lf\n", fv[sumprice]);
+	printf("fv[sumprod] = %Lf\n", fv[sumprod]);
+	printf("fv[sumsqkm] = %Lf\n", fv[sumsqkm]);
+	printf("fv[sumsqprice] = %Lf\n", fv[sumsqprice]);
 	printf("\n");
 	fv[final_b] = fv[tt1] * 10000.0;
 	fv[k_predict] = 10.0;
-	printf("fv[k_predict] = %f\n", fv[k_predict]);
-	printf("fv[learning_rate] = %f\n", fv[learning_rate]);
-	printf("fv[meank] = %f\n", fv[meank]);
-	printf("fv[meanp] = %f\n", fv[meanp]);
-	printf("fv[t0] = %f\n", fv[t0]);
-	printf("fv[t1] = %f\n", fv[t1]);
+	printf("fv[k_predict] = %Lf\n", fv[k_predict]);
+	printf("fv[learning_rate] = %Lf\n", fv[learning_rate]);
+	printf("fv[meank] = %Lf\n", fv[meank]);
+	printf("fv[meanp] = %Lf\n", fv[meanp]);
+	printf("fv[t0] = %Lf\n", fv[t0]);
+	printf("fv[t1] = %Lf\n", fv[t1]);
 	printf("\n");
 	/*ft_corr_rate(fv, fdb);
-	printf("fv[r] = %f\n", fv[r]);*/
+	printf("fv[r] = %Lf\n", fv[r]);*/
 	printf("\n");
 	ft_regr(fdb, fv);
-	/*printf("fv[costa] = %f\n", fv[costa]);
-	printf("fv[costb] = %f\n", fv[costb]);
-	printf("fv[costc] = %f\n", fv[costc]);
-	printf("fv[costd] = %f\n", fv[costd]);
-	printf("fv[final_b] = %f\n", fv[final_b]);
-	printf("fv[dtt0] = %f\n", fv[dtt0]);
-	printf("fv[dtt1] = %f\n", fv[dtt1]);
-	printf("fv[tt0] = %f\n", fv[tt0]);
-	printf("fv[tt1] = %f\n", fv[tt1]);
-	printf("lf[userkm] = %f\n", fv[userkm]);
-	printf("lf[userprice] = %f\n", fv[userprice]);*/
+	/*printf("fv[costa] = %Lf\n", fv[costa]);
+	printf("fv[costb] = %Lf\n", fv[costb]);
+	printf("fv[costc] = %Lf\n", fv[costc]);
+	printf("fv[costd] = %Lf\n", fv[costd]);
+	printf("fv[final_b] = %Lf\n", fv[final_b]);
+	printf("fv[dtt0] = %Lf\n", fv[dtt0]);
+	printf("fv[dtt1] = %Lf\n", fv[dtt1]);
+	printf("fv[tt0] = %Lf\n", fv[tt0]);
+	printf("fv[tt1] = %Lf\n", fv[tt1]);
+	printf("lf[userkm] = %Lf\n", fv[userkm]);
+	printf("lf[userprice] = %Lf\n", fv[userprice]);*/
 	//ft_userprice(fv);
 	return (0);
 }
